rename the preload folder to shard_runtime
