+++ luarules/gadgets/ai/boot.lua
@@ -1,11 +1,11 @@
 -- @TODO: Move top level logic out of ai.lua into here, and then make it load environment specific boot.lua files
 
 if ShardSpringLua then
-	-- VFS.Include( "luarules/gadgets/ai/preload/spring_lua/boot.lua" )
+	-- VFS.Include( "luarules/gadgets/ai/shard_runtime/spring_lua/boot.lua" )
 elseif game_engine then
 	require "spring_cpp/boot"
 else
-	require "preload/shard_null/boot"
+	require "shard_runtime/shard_null/boot"
 end
 
 shard_include( "ai" )
