--- luarules/gadgets/ai/preload/shard_null/boot.lua
@@ -1,16 +1,0 @@
-
-
-function shard_include( file )
-	if type(file) ~= 'string' then
-		return nil
-	end
-	return require( file )
-end
-
--- load null objects
-require "preload/hooks"
-require "preload/class"
-require "preload/aibase"
-require "preload/module"
-shard_include "preload/shard_null/unit"
-shard_include "preload/shard_null/unittype"
