+++ luarules/gadgets/ai/shard_runtime/spring_cpp/boot.lua
@@ -26,5 +26,5 @@ shard_include "class"
 shard_include "aibase"
 shard_include "module"
 
-shard_include "preload/spring_cpp/unit"
-shard_include "preload/spring_cpp/unittype"
+shard_include "shard_runtime/spring_cpp/unit"
+shard_include "shard_runtime/spring_cpp/unittype"
