+++ luarules/gadgets/ai/shard_runtime/spring_lua/fakeapi.lua
