+++ luarules/gadgets/ai/shard_runtime/spring_lua/fakeos.lua
